json_transformer_project/requirements.txt
boto3
python-dateutil
pytest
